refactor: move column adapters to a dedicated package
