+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/adapter/BotServiceAdapter.kt
@@ -1,8 +1,9 @@
-package com.wire.kalium.persistence.dao
+package com.wire.kalium.persistence.adapter
 
 import app.cash.sqldelight.ColumnAdapter
+import com.wire.kalium.persistence.dao.BotEntity
 
-class BotServiceAdapter : ColumnAdapter<BotEntity, String> {
+internal class BotServiceAdapter : ColumnAdapter<BotEntity, String> {
 
     override fun decode(databaseValue: String): BotEntity {
         val components = databaseValue.split("@")
