+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/adapter/ContentTypeAdapter.kt
@@ -1,14 +1,14 @@
-package com.wire.kalium.persistence.dao
+package com.wire.kalium.persistence.adapter
 
 import app.cash.sqldelight.ColumnAdapter
 import com.wire.kalium.persistence.dao.message.MessageEntity.ContentType
 import com.wire.kalium.persistence.dao.message.MessageEntity.ContentType.TEXT
 
-class ContentTypeAdapter : ColumnAdapter<ContentType, String> {
+internal class ContentTypeAdapter : ColumnAdapter<ContentType, String> {
 
     override fun decode(databaseValue: String): ContentType =
         ContentType.values().firstOrNull { it.name == databaseValue } ?: TEXT
 
     override fun encode(value: ContentType) = value.name
-    
+
 }
