+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/adapter/LogoutReasonAdapter.kt
@@ -1,9 +1,9 @@
-package com.wire.kalium.persistence.daokaliumdb
+package com.wire.kalium.persistence.adapter
 
 import app.cash.sqldelight.ColumnAdapter
 import com.wire.kalium.persistence.model.LogoutReason
 
-object LogoutReasonAdapter : ColumnAdapter<LogoutReason, String> {
+internal object LogoutReasonAdapter : ColumnAdapter<LogoutReason, String> {
     override fun decode(databaseValue: String): LogoutReason = LogoutReason.valueOf(databaseValue)
 
     override fun encode(value: LogoutReason): String = value.name
