+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/adapter/MemberRoleAdapter.kt
@@ -1,15 +1,16 @@
-package com.wire.kalium.persistence.dao
+package com.wire.kalium.persistence.adapter
 
 import app.cash.sqldelight.ColumnAdapter
+import com.wire.kalium.persistence.dao.Member
 
-class MemberRoleAdapter : ColumnAdapter<Member.Role, String> {
+internal class MemberRoleAdapter : ColumnAdapter<Member.Role, String> {
     override fun decode(databaseValue: String): Member.Role = when (databaseValue) {
         ADMIN -> Member.Role.Admin
         MEMBER -> Member.Role.Member
         else -> Member.Role.Unknown(databaseValue)
     }
 
-    override fun encode(value: Member.Role): String = when(value) {
+    override fun encode(value: Member.Role): String = when (value) {
         Member.Role.Admin -> ADMIN
         Member.Role.Member -> MEMBER
         is Member.Role.Unknown -> value.name
