+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/db/TableMapper.kt
@@ -23,13 +23,13 @@ import com.wire.kalium.persistence.Reaction
 import com.wire.kalium.persistence.Receipt
 import com.wire.kalium.persistence.SelfUser
 import com.wire.kalium.persistence.User
-import com.wire.kalium.persistence.dao.BotServiceAdapter
-import com.wire.kalium.persistence.dao.ContentTypeAdapter
-import com.wire.kalium.persistence.dao.ConversationAccessListAdapter
-import com.wire.kalium.persistence.dao.ConversationAccessRoleListAdapter
-import com.wire.kalium.persistence.dao.MemberRoleAdapter
-import com.wire.kalium.persistence.dao.QualifiedIDAdapter
-import com.wire.kalium.persistence.dao.QualifiedIDListAdapter
+import com.wire.kalium.persistence.adapter.BotServiceAdapter
+import com.wire.kalium.persistence.adapter.ContentTypeAdapter
+import com.wire.kalium.persistence.adapter.ConversationAccessListAdapter
+import com.wire.kalium.persistence.adapter.ConversationAccessRoleListAdapter
+import com.wire.kalium.persistence.adapter.MemberRoleAdapter
+import com.wire.kalium.persistence.adapter.QualifiedIDAdapter
+import com.wire.kalium.persistence.adapter.QualifiedIDListAdapter
 
 internal object TableMapper {
     val callAdapter = Call.Adapter(
